improved appearance of user display in users component and infouser and buttoninfo styled component throughout app; personalised css for text and buttons for mentioned styled components
